style: satisfy SwiftLint strict for new provider code
Move the iCloud-Keychain sync helpers into KeychainSync.swift (keeps
SecureStorage.swift byte-identical to main, matching the lint baseline), bundle
the audio-conversion tap context into a struct to cut the parameter count, and
add targeted (next/previous) SwiftLint suppressions for the naturally long
provider clients + settings screens. swift test: 541 pass; SwiftLint --strict
with baseline: 0 violations.
